bonus is 80%

+++ philo_bonus/philo_bonus.c
@@ -6,7 +6,7 @@
 /*   By: yel-ouaz <yel-ouaz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/21 20:44:34 by yel-ouaz          #+#    #+#             */
-/*   Updated: 2024/12/24 16:59:57 by yel-ouaz         ###   ########.fr       */
+/*   Updated: 2024/12/25 20:41:29 by yel-ouaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	wait_childreen(void)
 	int exit_s;
 
 	pid = waitpid(-1, &exit_s, 0);
-	while(pid != -1)
+	while (pid != -1)
 	{
 		pid = waitpid(-1, &exit_s, 0);
 	}
@@ -30,7 +30,7 @@ void kill_process(int *ids, int number)
 	int i;
 
 	i = 0;
-	while(i < number)
+	while (i < number)
 	{
 		kill(ids[i], SIGKILL);
 		i++;
@@ -52,13 +52,118 @@ void	*ft_calloc(int size)
 
 void	release_semaphores(t_data *sim_d)
 {
-	sem_close(sim_d->writing);
-	sem_close(sim_d->forks);
-	sem_unlink("/lektaba");
-	sem_unlink("/forks");
+	if (sim_d->writing)
+	{
+		sem_close(sim_d->writing);
+		sem_unlink(WRITESM);
+	}
+	if (sim_d->forks)
+	{
+		sem_close(sim_d->forks);
+		sem_unlink(FORKSM);
+	}
+	if (sim_d->dead)
+	{
+		sem_close(sim_d->dead);
+		sem_unlink(DEADSM);
+	}
+	if (sim_d->full)
+	{
+		sem_close(sim_d->full);
+		sem_unlink(FULLSM);
+	}
+	if (sim_d->philo_lock)
+	{
+		sem_close(sim_d->philo_lock);
+		sem_unlink(PHILOSM);
+	}
+	if (sim_d->green_light)
+	{
+		sem_close(sim_d->green_light);
+		sem_unlink(GREEN);
+	}
 }
 
-int	*start_sim(t_data *sim_d)
+void	*watcher1(void *arg)
+{
+	int			i;
+	t_watcher	*watcher_d;
+	
+	i = 0;
+	watcher_d = (t_watcher *)arg;
+	while (i < watcher_d->sim_d->num_philos)
+	{
+		sem_wait(watcher_d->sim_d->full);
+		i++;
+		if (watcher_d->sim_state == DEAD)
+			return (NULL);
+		if (i == watcher_d->sim_d->num_philos)
+		{
+			watcher_d->full = FULL;
+			sem_post(watcher_d->sim_d->dead);
+			return (NULL);
+		}
+	}
+	return (NULL);
+}
+
+void	*watcher2(void *arg)
+{
+	int i;
+	t_watcher *watcher_d;
+
+	i = 0;
+	watcher_d = (t_watcher *)arg;
+	sem_wait(watcher_d->sim_d->dead);
+	watcher_d->sim_state = DEAD;
+	while(i < watcher_d->sim_d->num_philos)
+	{
+		sem_post(watcher_d->sim_d->full);
+		i++;
+	}
+	return (NULL);
+}
+
+void	watchers(t_data *sim_d, int *ids)
+{
+	t_watcher watcher_d1;
+	t_watcher watcher_d2;
+	pthread_t watcher_th[2];
+
+	watcher_d1.sim_state = ALIVE;
+	watcher_d2.sim_state = ALIVE;
+	watcher_d1.full = HUNGRY;
+	watcher_d2.full = HUNGRY;
+	watcher_d1.sim_d = sim_d;
+	watcher_d2.sim_d = sim_d;
+	watcher_d1.ids = ids;
+	watcher_d2.ids = ids;
+	pthread_create(watcher_th, NULL, watcher1, &watcher_d1);
+	pthread_create(watcher_th + 1, NULL, watcher2, &watcher_d2);
+	pthread_join(watcher_th[0], NULL);
+	pthread_join(watcher_th[1], NULL);
+	kill_process(ids, sim_d->num_philos);
+	release_semaphores(sim_d);
+}
+
+// void	forking_fail(t_data *sim, int green_light)
+// {
+// 	if (green_light)
+// }
+
+void start_kids(t_data *sim_d)
+{
+	int i;
+
+	i = 0;
+	while(i <= sim_d->num_philos)
+	{
+		sem_post(sim_d->green_light);
+		i++;
+	}
+}
+
+int	*make_childreen(t_data *sim_d)
 {
 	int i;
 	int	f;
@@ -66,40 +171,46 @@ int	*start_sim(t_data *sim_d)
 	
 	i = 0;
 	ids = ft_calloc(sizeof(int) * sim_d->num_philos);
-	while(i < sim_d->num_philos)
+	if (!ids)
+	{
+		release_semaphores(sim_d);
+		exit(1);
+	}
+	sim_d->number_of_meals = 0;
+	sim_d->start = elapsed_time(0);
+	while (i < sim_d->num_philos)
 	{
 		sim_d->id = i + 1;
+		// sim_d->last_meal = elapsed_time(sim_d->start);
 		f = fork();
-		if ( f == -1)
+		if (f == -1)
 		{
 			kill_process(ids, i);
 			release_semaphores(sim_d);
-			exit(1);	
+			exit(1);
 		}
 		if (f == 0)
 			philo(sim_d);
 		else
 			ids[i] = f;
+		i++;
 	}
+	start_kids(sim_d);
 	return (ids);
-}
-
-
-void	watcher(t_data	*sim_d)
-{
-	
 }
 
 int main(int ac, char **av)
 {
+	int *ids;
 	t_data sim_d;
 
+	memset(&sim_d, 0, sizeof(sim_d));
 	if (check_args(ac, av))
 		return (1);
 	if (struct_init(ac, av, &sim_d))
 		return (2);
 	if (init_semaphore(&sim_d))
 		return (3);
-	start_sim(&sim_d);
-	watcher(&sim_d);
+	ids = make_childreen(&sim_d);
+	watchers(&sim_d, ids);
 }
